+++ internal/handlers/players/finish.go
@@ -51,6 +51,11 @@ func (h *PlayerHandler) Finish(w http.ResponseWriter, r *http.Request) {
 		http.Redirect(w, r, r.Header.Get("Referer"), http.StatusFound)
 		return
 	}
+	// If the user is in preview mode, only render the template, not the full layout.
+	template := templates.Start(*team, pageBlocks, blockStates)
+	if r.Context().Value(contextkeys.PreviewKey) == nil {
+		template = templates.Layout(template, "Start", team.Messages)
+	}
 
 	// data["notifications"], _ = h.NotificationService.GetNotifications(r.Context(), team.Code)
 	c := templates.Finish(*team, pageBlocks, blockStates)