+++ internal/handlers/players/start.go
@@ -30,7 +30,7 @@ func (h *PlayerHandler) Start(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	// Get blocks for the start page (lobby context)
+	// Get blocks for the start page
 	pageBlocks, blockStates, err := h.blockService.FindByOwnerIDAndTeamCodeWithStateAndContext(
 		r.Context(),
 		team.InstanceID,
@@ -38,20 +38,20 @@ func (h *PlayerHandler) Start(w http.ResponseWriter, r *http.Request) {
 		blocks.ContextStart,
 	)
 	if err != nil {
-		h.logger.Error("getting lobby blocks", "error", err.Error())
+		h.logger.Error("getting start blocks", "error", err.Error())
 		http.Redirect(w, r, r.Header.Get("Referer"), http.StatusFound)
 		return
 	}
 
 	// If the user is in preview mode, only render the template, not the full layout.
-	template := templates.Lobby(*team, pageBlocks, blockStates)
+	template := templates.Start(*team, pageBlocks, blockStates)
 	if r.Context().Value(contextkeys.PreviewKey) == nil {
 		template = templates.Layout(template, "Start", team.Messages)
 	}
 
 	err = template.Render(r.Context(), w)
 	if err != nil {
-		h.logger.Error("rendering lobby", "error", err.Error())
+		h.logger.Error("rendering start", "error", err.Error())
 	}
 }