+++ internal/middlewares/start_middleware.go
@@ -8,8 +8,8 @@ import (
 	"github.com/nathanhollows/Rapua/v6/models"
 )
 
-// LobbyMiddleware redirects to the start if the game is scheduled to start.
-func LobbyMiddleware(teamService teamService, next http.Handler) http.Handler {
+// StartMiddleware redirects to the start if the game is scheduled to start.
+func StartMiddleware(teamService teamService, next http.Handler) http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		// Preview requests should pass through
 		if r.Context().Value(contextkeys.PreviewKey) != nil {